+++ src/index.ts
@@ -3,6 +3,7 @@ import { ApplicationRoleEndpoint } from "./endpoints/applicationRole";
 import { AttachmentEndpoint } from "./endpoints/attachment";
 import { AvatarEndpoint } from "./endpoints/avatar";
 import { ComponentEndpoint } from "./endpoints/component";
+import { ConfigurationEndpoint } from "./endpoints/configuration";
 import { PermissionsEndpoint } from "./endpoints/permissions";
 import { Basic, BasicAuth } from "./types";
 
@@ -61,6 +62,11 @@ export class JiraServerConnector {
     component: ComponentEndpoint;
 
     /**
+     * Contains all operations from '/rest/api/latest/configuration/*'.
+     */
+    configuration: ConfigurationEndpoint;
+
+    /**
      * Instance new Jira Server Connector with user credentials and Jira host.
      * @param {BasicAuth} auth Basic Authorization info and Jira host 
      */
@@ -72,6 +78,7 @@ export class JiraServerConnector {
         this.attachments = new AttachmentEndpoint(this.auth);
         this.avatar = new AvatarEndpoint(this.auth);
         this.component = new ComponentEndpoint(this.auth);
+        this.configuration = new ConfigurationEndpoint(this.auth);
     }